Updated Full Code to the latest 2nd

+++ PBL3REAL/ViewModel/RoomTypeVM.cs
@@ -10,15 +10,14 @@ namespace HotelManagement.ViewModel
         public RoomTypeVM()
         {
             ListImg = new List<ImageVM>();
-            /* MapImgUrl = new Dictionary<int, string>();*/
+            //MapImgUrl = new Dictionary<int, string>();
         }
         public int IdRoomtype { get; set; }
         public string RotyName { get; set; }
         public string RotyDescription { get; set; }
-        public decimal RotyCurrentprice { get; set; }
+        public int RotyCurrentprice { get; set; }
         public int RotyCapacity { get; set; }
-
-        /* public IDictionary<int, string> MapImgUrl { get; set; }*/
+        //public IDictionary<int, string> MapImgUrl { get; set; }
         public List<ImageVM> ListImg { get; set; }
     }
 }
